+++ Src/Private/Add-DiaNodeFiller.ps1
@@ -1,4 +1,4 @@
-Function Add-DiaNodeFiller {
+function Add-DiaNodeFiller {
     <#
     .SYNOPSIS
         Function to create a html table used as Filler
@@ -32,6 +32,7 @@ Function Add-DiaNodeFiller {
             Mandatory = $false,
             HelpMessage = 'Enables debug mode for icons, highlighting the table in red.'
         )]
+        [Alias("DraftMode")]
         [bool] $IconDebug = $false,
 
         [Parameter(