+++ Src/Public/New-Diagrammer.ps1
@@ -108,6 +108,9 @@ function New-Diagrammer {
     .PARAMETER MainGraphBGColor
         Specifies the background color of the diagram. Default is 'White'.
 
+    .PARAMETER MainGraphSize
+        Specifies the image resolution size. (e.g., 8,11! = 800x1100 pixels) Default = None.
+
     .NOTES
         Version:        0.2.24
         Author(s):      Jonathan Colon
@@ -211,7 +214,7 @@ function New-Diagrammer {
                     throw "Path $_ not found!"
                 }
             })]
-        [System.IO.FileInfo] $IconPath,
+        [System.IO.FileInfo] $IconPath = (Join-Path (Split-Path (Split-Path $PSScriptRoot -Parent) -Parent) 'icons'),
 
         [Parameter(
             Position = 4,
@@ -219,7 +222,10 @@ function New-Diagrammer {
             HelpMessage = 'Please provide the icons hashtable'
         )]
         [ValidateNotNullOrEmpty()]
-        [Hashtable] $ImagesObj,
+        [Hashtable] $ImagesObj = @{
+            "Main_Logo" = "Diagrammer.png"
+            "Logo_Footer" = "Diagrammer_footer.png"
+        },
 
         [Parameter(
             Mandatory = $false,
@@ -385,9 +391,14 @@ function New-Diagrammer {
             Mandatory = $false,
             HelpMessage = 'Allow to set diagram backgroud color'
         )]
-        [string] $MainGraphBGColor = 'White'
-    )
+        [string] $MainGraphBGColor = 'White',
 
+        [Parameter(
+            Mandatory = $false,
+            HelpMessage = 'Allow to set image resolution size (Ex: 8,11! = 800x1100 pixels)'
+        )]
+        [string] $MainGraphSize
+    )
 
     begin {
 
@@ -397,16 +408,6 @@ function New-Diagrammer {
         } else {
             $global:VerbosePreference = 'SilentlyContinue'
             $global:DebugPreference = 'SilentlyContinue'
-        }
-
-        # Variable translating Icon to Image Path ($IconPath)
-        if ($ImagesObj) {
-            $script:Images = $ImagesObj
-        } else {
-            $script:Images = @{
-                "Main_Logo" = "Diagrammer.png"
-                "Logo_Footer" = "Diagrammer_footer.png"
-            }
         }
 
         if (($Format -ne "base64") -and !(Test-Path $OutputFolderPath)) {
@@ -439,16 +440,16 @@ function New-Diagrammer {
         }
 
         # Validate Custom logo
-        if ($Logo -and (-Not $LogoName)) {
-            $CustomLogo = Test-Logo -LogoPath (Get-ChildItem -Path $Logo).FullName -IconPath $IconPath -ImagesObj $Images
+        if ($Logo -and (-not $LogoName)) {
+            $CustomLogo = Test-Logo -LogoPath (Get-ChildItem -Path $Logo).FullName -IconPath $IconPath -ImagesObj $ImagesObj
         } elseif ($LogoName) {
             $CustomLogo = $LogoName
         } else {
             $CustomLogo = "Diagrammer.png"
         }
         # Validate Custom Signature Logo
-        if ($SignatureLogo -and (-Not $SignatureLogoName )) {
-            $CustomSignatureLogo = Test-Logo -LogoPath (Get-ChildItem -Path $SignatureLogo).FullName -IconPath $IconPath -ImagesObj $Images
+        if ($SignatureLogo -and (-not $SignatureLogoName )) {
+            $CustomSignatureLogo = Test-Logo -LogoPath (Get-ChildItem -Path $SignatureLogo).FullName -IconPath $IconPath -ImagesObj $ImagesObj
         } elseif ($SignatureLogoName) {
             $CustomSignatureLogo = $SignatureLogoName
         } else {
@@ -469,6 +470,10 @@ function New-Diagrammer {
             nodesep = $NodeSeparation
             ranksep = $SectionSeparation
             bgcolor = $MainGraphBGColor
+        }
+
+        if ($MainGraphSize) {
+            $MainGraphAttributes['size'] = $MainGraphSize
         }
     }
 
@@ -502,9 +507,9 @@ function New-Diagrammer {
             if ($Signature) {
                 Write-Verbose -Message "Generating diagram signature"
                 if ($CustomSignatureLogo) {
-                    $Signature = (Add-DiaHtmlSignatureTable -ImagesObj $Images -Rows "Author: $($AuthorName)", "Company: $($CompanyName)" -TableBorder 2 -CellBorder 0 -Align 'left' -Logo $CustomSignatureLogo -IconDebug $IconDebug)
+                    $Signature = (Add-DiaHtmlSignatureTable -ImagesObj $ImagesObj -Rows "Author: $($AuthorName)", "Company: $($CompanyName)" -TableBorder 2 -CellBorder 0 -Align 'left' -Logo $CustomSignatureLogo -IconDebug $IconDebug)
                 } else {
-                    $Signature = (Add-DiaHtmlSignatureTable -ImagesObj $Images -Rows "Author: $($AuthorName)", "Company: $($CompanyName)" -TableBorder 2 -CellBorder 0 -Align 'left' -Logo "Logo_Footer" -IconDebug $IconDebug)
+                    $Signature = (Add-DiaHtmlSignatureTable -ImagesObj $ImagesObj -Rows "Author: $($AuthorName)", "Company: $($CompanyName)" -TableBorder 2 -CellBorder 0 -Align 'left' -Logo "Logo_Footer" -IconDebug $IconDebug)
                 }
             } else {
                 Write-Verbose -Message "No diagram signature specified"
@@ -522,7 +527,7 @@ function New-Diagrammer {
             # Subgraph OUTERDRAWBOARD1 used to draw the footer signature (bottom-right corner)
             SubGraph OUTERDRAWBOARD1 -Attributes @{Label = $Signature; fontsize = 24; penwidth = 1.5; labelloc = 'b'; labeljust = "r"; style = $SubGraphDebug.style; color = $SubGraphDebug.color } {
                 # Subgraph MainGraph used to draw the main drawboard.
-                SubGraph MainGraph -Attributes @{Label = (Add-DiaHTMLLabel -ImagesObj $Images -Label $MainDiagramLabel -IconType $CustomLogo -IconDebug $IconDebug -IconWidth 250 -IconHeight 80 -Fontsize $MainDiagramLabelFontsize -fontColor  $MainDiagramLabelFontColor -fontName  $MainDiagramLabelFontname); fontsize = 24; penwidth = 0; labelloc = 't'; labeljust = "c" } {
+                SubGraph MainGraph -Attributes @{Label = (Add-DiaHTMLLabel -ImagesObj $ImagesObj -Label $MainDiagramLabel -IconType $CustomLogo -IconDebug $IconDebug -IconWidth 250 -IconHeight 80 -Fontsize $MainDiagramLabelFontsize -fontColor  $MainDiagramLabelFontColor -fontName  $MainDiagramLabelFontname); fontsize = 24; penwidth = 0; labelloc = 't'; labeljust = "c" } {
                     $InputObject
                 }
             }